requirements.txt
colorama==0.4.4
requests==2.26.0
python-nmap==0.6.1
shodan==1.26.0
